+++ framework/ops.py
@@ -2,12 +2,20 @@ from framework import tensor
 from framework.core import ops
 
 
+class OpRecord:
+    def __init__(self, func, func_backward, input_tensors, output_tensor):
+        self.func = func
+        self.func_backward = func_backward
+        self.input_tensors = input_tensors
+        self.output_tensor = output_tensor
+
+
 # Forward operations
 def matmul(a, b):
     result = tensor.Tensor(shape=(a.shape[0], b.shape[1]))
     ops.matmul(a.data, b.data, result.data)
 
-    result.op_record = tensor.OpRecord(
+    result.op_record = OpRecord(
         func=matmul,
         func_backward=matmul_backward,
         input_tensors=(a, b),
@@ -20,7 +28,7 @@ def add(a, b):
     result = tensor.Tensor(shape=a.shape)
     ops.add(a.data, b.data, result.data)
 
-    result.op_record = tensor.OpRecord(
+    result.op_record = OpRecord(
         func=add,
         func_backward=add_backward,
         input_tensors=(a, b),
@@ -33,7 +41,7 @@ def multiply(a, b):
     result = tensor.Tensor(shape=a.shape)
     ops.multiply(a.data, b.data, result.data)
 
-    result.op_record = tensor.OpRecord(
+    result.op_record = OpRecord(
         func=multiply,
         func_backward=multiply_backward,
         input_tensors=(a, b),
@@ -46,7 +54,7 @@ def relu(a):
     result = tensor.Tensor(shape=a.shape)
     ops.relu(a.data, result.data)
 
-    result.op_record = tensor.OpRecord(
+    result.op_record = OpRecord(
         func=relu,
         func_backward=relu_backward,
         input_tensors=(a,),
@@ -59,7 +67,7 @@ def softmax(a):
     result = tensor.Tensor(shape=a.shape)
     ops.softmax(a.data, result.data)
 
-    result.op_record = tensor.OpRecord(
+    result.op_record = OpRecord(
         func=softmax,
         func_backward=softmax_backward,
         input_tensors=(a,),
@@ -72,7 +80,7 @@ def log(a):
     result = tensor.Tensor(shape=a.shape)
     ops.log(a.data, result.data)
 
-    result.op_record = tensor.OpRecord(
+    result.op_record = OpRecord(
         func=log,
         func_backward=log_backward,
         input_tensors=(a,),
@@ -85,7 +93,7 @@ def sum(a):
     result = tensor.Tensor(shape=(1,))
     ops.sum(a.data, result.data)
 
-    result.op_record = tensor.OpRecord(
+    result.op_record = OpRecord(
         func=sum,
         func_backward=sum_backward,
         input_tensors=(a,),
